Don't use float32 as an internal takum representation
While it's true that one can express the logarithmic value of takum8
in float16 and that of takum16 in float32, the same does not hold for
the respective represented value due to insufficient dynamic range.

Thus we need to compute with float64 even in the case of Takum8 and
takum16.

We use this opportunity to improve the display of takums in the REPL by
roughly implementing the precision function for the worst case as a
rough reference for the number of significant digits to print. The use
of AnyTakum instead of distinct functions for each takum type helps drop
a lot of code.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>