Add instance-based precision()
Given takum is a tapered precision format the precision varies
depending on the _instance_, e.g. 1.0 has a higher precision
than 100.0 given it has more mantissa bits.

Julia implements the precision() function, e.g. precision(Float16),
and just calls into the type when you call precision with an instance
of Float16. However, we will do it the other way round:

We make use of the precision() function added to libtakum in version
0.2.0 and return the exact number of mantissa bits for a given
takum instance. If called with the type, we return the precision
of 0.0, which has been implemented to return the worst-case
precision max(0, n-12) of the format.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>