Make linear takums the default type
Takum{8,16,32,64} is now the floating-point type, whereas
LogTakum{8,16,32,64} is the logarithmic number system. This reflects
the fact that at the current point of research, as floating-point
formats are ubiquitous, it brought some confusion as to which type to
use.

A general user would expect Takum16, for instance, to be a
floating-point type. If a user is advanced enough to know about LNS,
they will be able to explicitly specify LogTakum16 and know what it
entails.

This also brings Takums.jl in line with Posits.jl.

Signed-off-by: Laslo Hunhold <hunhold@uni-koeln.de>